requirements/base.txt
Django==1.5.5
South==0.8.2
psycopg2==2.5.1
django-braces==1.2.2
django-registration==1.0
django-crispy-forms==1.3.2
crispy-forms-foundation==0.2.1
django-model-utils==1.5.0
django-autoslug==1.7.1